Fixes to work with current API and OS
Open Sprinkler API has changed, causing errors in opsiLCD.py. Updated
to work with v.2.1.9.

Also, fixed the network IP detection. Previous method of "hostname -I"
returned all addresses, including IPv6, with no guaranteed order. Now
uses the netifaces library and a defined variable for specifying which
interface will be used (wlan, eth0, etc.) The downside is that it
requires installation of netifaces. The upside is that it should be
a bit more future-proof. Here's hoping.