First MQTT enabled version with new features
Added MQTT subscription notifications to avoid the need to poll the
OpenSprinkler system for status. This essentially provides a "push"
notification from  the OpenSprinkler system to this script.

Secondly, this changes the model of running the script from a
"one shot" script that needs to be run from a repeating cron job to an
ongoing service.

On a more minor note, this also added the ability to auto dim the LCD
backlight. It will automatically light up on a new status update, and
then dim after a configurable number of seconds.